modification de la classe Matrice et ajout du début de la classe Repere

+++ src/main/java/utilitaire/Matrice.java
@@ -1,9 +1,9 @@
 package utilitaire;
 
 public class Matrice {
-	public static int[][] addition(int[][] m1, int[][] m2) throws Exception{
+	public static double[][] addition(double[][] m1, double[][] m2) throws Exception{
 		if(m1.length==m2.length && m1[0].length==m2[0].length) {
-			int[][] res=new int[m1.length][m1[0].length];
+			double[][] res=new double[m1.length][m1[0].length];
 			for(int i=0;i<res.length;i++) {
 				for(int j=0;j<res[0].length;j++) {
 					res[i][j]=m1[i][j]+m2[i][j];
@@ -15,9 +15,9 @@ public class Matrice {
 		}
 		
 	}
-	public static int[][] multiplication(int[][] m1, int[][] m2) throws Exception{
+	public static double[][] multiplication(double[][] m1, double[][] m2) throws Exception{
 		if(m1[0].length==m2.length) {
-			int[][] res=new int[m1.length][m2[0].length];
+			double[][] res=new double[m1.length][m2[0].length];
 			for(int i=0;i<res.length;i++) {
 				for(int j=0;j<res[0].length;j++) {
 					res[i][j]=additionLigneColonne(m1, m2, i, j);
@@ -29,8 +29,8 @@ public class Matrice {
 		}
 		
 	}
-	public static int additionLigneColonne(int[][] m1, int[][] m2, int lig, int col) {
-		int res=0;
+	public static double additionLigneColonne(double[][] m1, double[][] m2, int lig, int col) {
+		double res=0;
 		for(int i=0; i<m1[0].length;i++) {
 				res+=m1[lig][i]*m2[i][col];
 		}
